% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Function.R
\name{create_sir_plot}
\alias{create_sir_plot}
\title{Plotting function for SIR model simulation results}
\usage{
create_sir_plot(output_df, title = "SIR Model Simulation")
}
\arguments{
\item{output_df}{A data frame containing the simulation results with columns 'time', 'S', 'I', 'R'.}

\item{title}{A character string specifying the title of the plot. Default is "SIR Model Simulation".}
}
\value{
A ggplot object representing the SIR model simulation plot.
}
\description{
This function creates a line plot to visualize the simulation results of the SIR (Susceptible-Infectious-Recovered) model.
}
\examples{
# Example usage:
# Assuming sir_result_object@output is a data frame with columns 'time', 'S', 'I', 'R'
sir_result_object <- sir_result_object@output
sir_plot <- create_sir_plot(sir_result_object)
print(sir_plot)

}
\author{
Joao Filipe Siun Mulcahy Ellen Creed
}
