% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_function.R
\docType{data}
\name{initial_state}
\alias{initial_state}
\title{Defines a class to represent the results of the SIR model simulation.}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
initial_state
}
\arguments{
\item{.Object}{An object of class SIRResult.}

\item{initial_state}{Numeric vector representing the initial state of the system.}

\item{times}{Numeric vector representing the time points for simulation.}

\item{sir_model}{A function defining the SIR model dynamics.}

\item{parameters}{Numeric vector representing the parameters of the SIR model.}

\item{object}{An object of class SIRResult.}
}
\value{
An object of class SIRResult.

An updated object of class SIRResult with the simulation results.
}
\description{
Represents the results of the SIR model simulation.

Initializes an object of class SIRResult.

Generates SIR model simulation results based on the specified parameters.
}
\details{
SIRResult Class Definition
}
\examples{
# Define initial state and parameters
initial_state <- c(S = 0.99, I = 0.01, R = 0)
parameters <- c(beta = 0.3, gamma = 0.1)

# Define time points
times <- seq(0, 80, by = 1)

# Create an object of class SIRResult
sir_result_object <- new("SIRResult",
                         initial_state = initial_state,
                         times = times,
                         sir_model = function(time, state, parameters) {
                           with(as.list(c(state, parameters)), {
                             dS <- -beta * S * I
                             dI <- beta * S * I - gamma * I
                             dR <- gamma * I
                             return(list(c(dS, dI, dR)))
                           })
                         },
                         parameters = parameters)

# Generate the SIR results
sir_result_object <- generate(sir_result_object)

# View the output
print(sir_result_object@output)
}
\seealso{
\code{\link{initialize.SIRResult}}, \code{\link{generate.SIRResult}}
Constructor method for the SIRResult class
}
\author{
Joao Filipe Siun Mulcahy Ellen Creed

Joao Filipe Siun Mulcahy Ellen Creed
Generate SIR results
}
\keyword{"SIRResult"}
\keyword{"class"}
\keyword{SIRResult}
\keyword{generate}
\keyword{initialize}
\keyword{methods}
